pyproject.toml
[build-system]
requires = ["setuptools ~= 75.1"]
build-backend = "setuptools.build_meta"

[project]
name = "gurobi_sphinxtheme"
version = "1.0.0"
description = "Gurobi Sphinx theme"
readme = "README.md"
dependencies = [
    "furo",
]

[project.entry-points."sphinx.html_themes"]
gurobi_sphinxtheme = "gurobi_sphinxtheme"
docs_gurobi_com = "docs_gurobi_com"

[project.optional-dependencies]
docs_gurobi_com = [
    "beautifulsoup4",
    "sphinx_sitemap>=2.6,<2.7",
    "furo == 2024.1.29",
]

[tool.setuptools]
packages = [
    "gurobi_sphinxtheme",
    "docs_gurobi_com",
]

[tool.setuptools.package-data]
gurobi_sphinxtheme = [
    "theme/theme.conf",
    "theme/*.html",
    "theme/static/*.svg",
    "theme/static/*.css",
    "theme/static/*.js",
]
docs_gurobi_com = [
    "theme/theme.conf",
    "theme/*.html",
    "theme/static/*.svg",
    "theme/static/*.css",
    "theme/static/*.js",
    "theme/static/*.ico",
    "latex/*",
]
